Fixing error in casting of double math and tests associated to it

+++ rskj-core/src/main/java/org/ethereum/cost/InitcodeCostCalculator.java
@@ -40,7 +40,7 @@ public class InitcodeCostCalculator implements CostCalculator{
     @Override
     public long calculateCost(int dataLength, ActivationConfig.ForBlock activations) {
         if( activations.isActive(ConsensusRule.RSKIP438) ) {
-            return  INITCODE_WORD_COST  *  (long) Math.ceil(dataLength / 32);
+            return  INITCODE_WORD_COST  *  (long) Math.ceil((double)dataLength / 32);
         }
         return 0;
     }
